Adding foreign keys Changing key types from bigint to int
git-tfs-id: [https://tfs.codeplex.com/tfs/tfs04]$/polaris;C38300

+++ Main/Polaris.Dbl/Schema Objects/Tables/Polaris.DevelopmentTeam.table.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [Polaris].[DevelopmentTeam]
 (
-[DevelopmentTeamId] [bigint] NOT NULL IDENTITY(1, 1),
+[DevelopmentTeamId] [int] NOT NULL IDENTITY(1, 1),
 [Name] [varchar] (max) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
 [Key] [uniqueidentifier] NOT NULL
 ) ON [PRIMARY] TEXTIMAGE_ON [PRIMARY]
